Release 5.4.0 * Now requires C++14. * New Samples   * GLES2.TextureCompression   * Vulkan.ObjectSelection   * Vulkan.SRGBFramebuffer * Vulkan screenshot support * Replaced TCLAP with a custom argument parser to reduce executable size. * FslBuild   * now supports “-c install”.   * Experimental CMake support (see FslBuild_CMake.md)   * Experimental VS2019 support   * Android Studio 3.5.0   * Android NDK r20

+++ DemoFramework/FslDemoHost/Base/source/FslDemoHost/Base/Service/Mouse/MouseService.cpp
@@ -110,7 +110,7 @@ namespace Fsl
     else
     {
       // Check if the button was pressed before
-      if ((m_buttonState.Flags & button) == button)
+      if ((m_buttonState.Flags & static_cast<uint32_t>(button)) == static_cast<uint32_t>(button))
       {
         m_buttonState.Flags &= ~button;
